Moves finalizeGetTileOrGrid to mapcontroller it's no longer used from any other place
