+++ lib/windshaft/controllers/map.js
@@ -116,24 +116,22 @@ MapController.prototype.layer = function(req, res) {
 };
 
 MapController.prototype.tileOrLayer = function (req, res) {
-    this._app.doCORS(res);
-    return getTileOrGrid(this._app, this._mapBackend, req, res);
-};
+    var self = this;
 
-function getTileOrGrid(app, mapBackend, req, res) {
+    this._app.doCORS(res);
     step(
         function() {
-            app.req2params(req, this);
+            self._app.req2params(req, this);
         },
         function(err) {
             req.profiler.done('req2params');
             if ( err ) {
                 throw err;
             }
-            mapBackend.getTileOrGrid(req, res, this);
+            self._mapBackend.getTileOrGrid(req, res, this);
         },
         function finalize(err, req_ret, res_ret, tile, headers) {
-            app.finalizeGetTileOrGrid(err, req, res, tile, headers);
+            self.finalizeGetTileOrGrid(err, req, res, tile, headers);
             return null;
         },
         function finish(err) {
@@ -142,6 +140,43 @@ function getTileOrGrid(app, mapBackend, req, res) {
             }
         }
     );
-}
+};
 
-module.exports.getTileOrGrid = getTileOrGrid;
+// This function is meant for being called as the very last
+// step by all endpoints serving tiles or grids
+MapController.prototype.finalizeGetTileOrGrid = function(err, req, res, tile, headers) {
+    var supportedFormats = {
+        grid_json: true,
+        json_torque: true,
+        torque_json: true,
+        png: true
+    };
+
+    var formatStat = 'invalid';
+    if (req.params.format) {
+        var format = req.params.format.replace('.', '_');
+        if (supportedFormats[format]) {
+            formatStat = format;
+        }
+    }
+
+    if (err){
+        // See https://github.com/Vizzuality/Windshaft-cartodb/issues/68
+        var errMsg = err.message ? ( '' + err.message ) : ( '' + err );
+        var statusCode = this._app.findStatusCode(err);
+
+        // Rewrite mapnik parsing errors to start with layer number
+        var matches = errMsg.match("(.*) in style 'layer([0-9]+)'");
+        if (matches) {
+            errMsg = 'style'+matches[2]+': ' + matches[1];
+        }
+
+        this._app.sendError(res, {error: errMsg}, statusCode, 'TILE RENDER', err);
+        global.statsClient.increment('windshaft.tiles.error');
+        global.statsClient.increment('windshaft.tiles.' + formatStat + '.error');
+    } else {
+        this._app.sendWithHeaders(res, tile, 200, headers);
+        global.statsClient.increment('windshaft.tiles.success');
+        global.statsClient.increment('windshaft.tiles.' + formatStat + '.success');
+    }
+};
