+++ lib/windshaft/server.js
@@ -350,49 +350,6 @@ module.exports = function(opts){
       app.sendResponse(res, [err, statusCode]);
     };
 
-    // This function is meant for being called as the very last
-    // step by all endpoints serving tiles or grids
-    app.finalizeGetTileOrGrid = function(err, req, res, tile, headers) {
-        var supportedFormats = {
-            grid_json: true,
-            json_torque: true,
-            png: true
-        };
-
-        var formatStat;
-        if (req.params.format) {
-            var format = req.params.format.replace('.', '_');
-            if (supportedFormats[format]) {
-                formatStat = format;
-            }
-        }
-
-      if (err){
-          // See https://github.com/Vizzuality/Windshaft-cartodb/issues/68
-          var errMsg = err.message ? ( '' + err.message ) : ( '' + err );
-          var statusCode = app.findStatusCode(err);
-
-          // Rewrite mapnik parsing errors to start with layer number
-          var matches = errMsg.match("(.*) in style 'layer([0-9]+)'");
-          if (matches) {
-            errMsg = 'style'+matches[2]+': ' + matches[1];
-          }
-
-          app.sendError(res, {error: errMsg}, statusCode, 'TILE RENDER', err);
-          statsClient.increment('windshaft.tiles.error');
-            if (formatStat) {
-                statsClient.increment('windshaft.tiles.' + formatStat + '.error');
-            }
-      } else {
-          app.sendWithHeaders(res, tile, 200, headers);
-          statsClient.increment('windshaft.tiles.success');
-            if (formatStat) {
-                statsClient.increment('windshaft.tiles.' + formatStat + '.success');
-            }
-      }
-
-    };
-
     var MapBackend = require('./backends/map');
     var mapBackend = new MapBackend(app, render_cache, map_store);
